requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.5.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
requests>=2.31.0
aiofiles>=23.1.0
jinja2>=3.1.2
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0
